Added expected_products_sensor to the _working method
Added expected_products_sensor to the _working method. The sensor has the responsibility to monitor the output level of the cell and to raise a flag everytime a piece is not placed in the expected moment.

The method has to be integrate in the _working() method.

Other minor changes has been made.